cleaning

+++ torch_geometric/nn/nlp/txt2kg.py
@@ -50,7 +50,7 @@ class TXT2KG():
             if not self.initd_LM:
                 from torch_geometric.nn.nlp import LLM
                 LM_name = "VAGOsolutions/SauerkrautLM-v2-14b-DPO"
-                self.model = LLM(LM_name, num_params=14, dtype=torch.float8_e5m2).eval()
+                self.model = LLM(LM_name, num_params=14).eval()
                 self.initd_LM = True
             out_strs = self.model.inference(
                 question=[txt + '\n' + self.system_prompt],
